05_FirstStepsInJavawithSpringBoot/rest-with-spring-boot-and-java-delfino/src/main/java/br/com/delfino/services/BookServices.java
package br.com.delfino.services;

import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;

import java.util.List;
import java.util.logging.Logger;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import br.com.delfino.BookController;
import br.com.delfino.data.vo.v1.BookVO;
import br.com.delfino.exceptions.RequiredObjectIsNullException;
import br.com.delfino.exceptions.ResourceNotFoundException;
import br.com.delfino.mapper.DozerMapper;
import br.com.delfino.mapper.custom.PersonMapper;
import br.com.delfino.model.Book;
import br.com.delfino.repositories.BookRepository;

@Service
public class BookServices {

	private Logger logger = Logger.getLogger(BookServices.class.getName());

	@Autowired
	BookRepository repository;

	@Autowired
	PersonMapper mapper;

	public List<BookVO> findAll() {
		logger.info("Finding all books...");

		var books = DozerMapper.parseListObjects(repository.findAll(), BookVO.class);

		books
			.stream()
			.forEach(b -> b.add(linkTo(methodOn(BookController.class).findById(b.getKey())).withSelfRel()));
		return books;
	}

	public BookVO findById(Long id) {
		logger.info("Finding book...");
		logger.info("id " + id);

		var entity = repository.findById(id)
				.orElseThrow(() -> new ResourceNotFoundException("No records found for this ID!"));
		
		var vo = DozerMapper.parseObject(entity, BookVO.class);
		vo.add(linkTo(methodOn(BookController.class).findById(id)).withSelfRel());
		return vo;
	}

	public BookVO create(BookVO book) {		
		if (book == null)
			throw new RequiredObjectIsNullException();
		
		logger.info("Creating book...");
		var entity = DozerMapper.parseObject(book, Book.class);
		var vo = DozerMapper.parseObject(repository.save(entity), BookVO.class);
		vo.add(linkTo(methodOn(BookController.class).findById(vo.getKey())).withSelfRel());
		return vo;

	}

	public BookVO update(BookVO book) {

		logger.info("Updating book...");

		if (book == null)
			throw new RequiredObjectIsNullException();

		var entity = repository.findById(book.getKey())
				.orElseThrow(() -> new ResourceNotFoundException("No records found for this ID!"));

		entity.setTitle(book.getTitle());
		entity.setAuthor(book.getAuthor());
		entity.setLaunchDate(book.getLaunchDate());
		entity.setPrice(book.getPrice());

		var vo = DozerMapper.parseObject(repository.save(entity), BookVO.class);
		vo.add(linkTo(methodOn(BookController.class).findById(vo.getKey())).withSelfRel());
		return vo;
	}

	public void delete(Long id) {
		logger.info("Deleting book...");

		var entity = repository.findById(id)
				.orElseThrow(() -> new ResourceNotFoundException("No records found for this ID!"));

		repository.delete(entity);
	}

}
